+++ examples/easycat/easycat_example.cc
@@ -1,6 +1,7 @@
-#include "kickcat/LinkSingle.h"
+#include "kickcat/Link.h"
 #include "kickcat/Bus.h"
 #include "kickcat/Prints.h"
+#include "kickcat/SocketNull.h"
 
 #ifdef __linux__
     #include "kickcat/OS/Linux/Socket.h"
@@ -18,15 +19,50 @@ using namespace kickcat;
 
 int main(int argc, char* argv[])
 {
-    if (argc != 2)
+    if (argc != 3 and argc != 2)
     {
-        printf("usage: ./test NIC\n");
+        printf("usage redundancy mode : ./test NIC_nominal NIC_redundancy\n");
+        printf("usage no redundancy mode : ./test NIC_nominal\n");
         return 1;
     }
 
-    auto socket = std::make_shared<Socket>();
 
-    std::shared_ptr<LinkSingle> link = std::make_shared<LinkSingle>(socket);
+    auto reportRedundancy = []()
+    {
+        printf("Redundancy has been activated due to loss of a cable \n");
+    };
+
+    auto socketNominal = std::make_shared<Socket>();
+
+    std::shared_ptr<AbstractSocket> socketRedundancy;
+    std::string red_interface_name;
+
+    if (argc == 2)
+    {
+        printf("No redundancy mode selected \n");
+        socketRedundancy = std::make_shared<SocketNull>();
+        red_interface_name = "null";
+    }
+    else
+    {
+        socketRedundancy = std::make_shared<Socket>();
+        red_interface_name = argv[2];
+    }
+
+    try
+    {
+        socketNominal->open(argv[1], 2ms);
+        socketRedundancy->open(red_interface_name, 2ms);
+    }
+    catch (std::exception const& e)
+    {
+        std::cerr << e.what() << std::endl;
+        return 1;
+    }
+
+    std::shared_ptr<Link> link= std::make_shared<Link>(socketNominal, socketRedundancy, reportRedundancy);
+    link->checkRedundancyNeeded();
+
     Bus bus(link);
 
     auto print_current_state = [&]()
@@ -41,9 +77,9 @@ int main(int argc, char* argv[])
     uint8_t io_buffer[2048];
     try
     {
-        socket->open(argv[1], 2ms);
         bus.init();
 
+        printf("Init done \n");
         print_current_state();
 
         bus.createMapping(io_buffer);
@@ -92,11 +128,11 @@ int main(int argc, char* argv[])
         return 1;
     }
 
-    socket->setTimeout(500us);
+    microseconds nominal_timeout = 500us;
+    socketNominal->setTimeout(nominal_timeout);
+    socketRedundancy->setTimeout(nominal_timeout);
 
     constexpr int64_t LOOP_NUMBER = 12 * 3600 * 1000; // 12h
-    FILE* stat_file = fopen("stats.csv", "w");
-    fwrite("latency\n", 1, 8, stat_file);
 
     auto& easycat = bus.slaves().at(0);
     int64_t last_error = 0;
@@ -106,7 +142,6 @@ int main(int argc, char* argv[])
 
         try
         {
-            nanoseconds t1 = since_epoch();
             bus.sendLogicalRead(callback_error);
             bus.sendLogicalWrite(callback_error);
             bus.sendRefreshErrorCounters(callback_error);
@@ -115,12 +150,10 @@ int main(int argc, char* argv[])
             bus.sendReadMessages(callback_error);
             bus.sendWriteMessages(callback_error);
             bus.finalizeDatagrams();
-            nanoseconds t2 = since_epoch();
 
 
-            nanoseconds t3 = since_epoch();
             bus.processAwaitingFrames();
-            nanoseconds t4 = since_epoch();
+
 
             for (int32_t j = 0;  j < easycat.input.bsize; ++j)
             {
@@ -128,25 +161,10 @@ int main(int argc, char* argv[])
             }
             printf("\r");
 
-            // blink a led - EasyCAT example for Arduino
-            if ((i % 50) < 25)
-            {
-                easycat.output.data[0] = 1;
-            }
-            else
-            {
-                easycat.output.data[0] = 0;
-            }
-
             if ((i % 1000) == 0)
             {
                 printErrorCounters(easycat);
             }
-
-            microseconds sample = duration_cast<microseconds>(t4 - t3 + t2 - t1);
-            std::string sample_str = std::to_string(sample.count());
-            fwrite(sample_str.data(), 1, sample_str.size(), stat_file);
-            fwrite("\n", 1, 1, stat_file);
         }
         catch (std::exception const& e)
         {
@@ -155,8 +173,6 @@ int main(int argc, char* argv[])
             std::cerr << e.what() << " at " << i << " delta: " << delta << std::endl;
         }
     }
-
-    fclose(stat_file);
 
     return 0;
 }
